Mir integration: Switch around Wayland client associations.
On the server side wl_client_create creates the wl_client immediately, not when the client
end of the socketpair is passed to wl_display_connect_to_fd().

This means that setting the socket associated with the *next* client to be created is
racy.

Instead, do the reverse; wait for the server-side client to be created, and then
associate it with the fd.