ClientSurfaceEventsTest.frame_timestamp_increases: Fix *again*.
This test turns out to be flaky; if you run it in a loop, about half the time it will
pass and about half the time it will fail (on my system).

This is because we weren't ensuring that the first `frame`+`commit` request was
*actually sent to the compositor* before we waited and then sent the second
`frame`+`commit` request.

Fix this by adding a `client.flush()` after the first `commit`.